Added: Enemy Data

+++ Assets/_Scripts/Data/Hero/HeroData.cs
@@ -8,8 +8,8 @@ public struct HeroData
 {
     public int _heroID;
     public string _name;
-    public int _maxHealth;
-    public int _attackDamage;
+    public float _maxHealth;
+    public float _attackDamage;
     public int _experience;
     public int _level;
     public Sprite _sprite;
